aiagents_news/pyproject.toml
[project]
name = "aiagent-news"
version = "0.1.0"
description = "Blog Writer Agent using OpenAI"
authors = [
    {name = "vapmail16",email = "vapmail16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"  
dependencies = [
    "openai (>=1.68.2,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "langchain (>=0.3.21,<0.4.0)",
    "praw (>=7.8.1,<8.0.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "langchain-openai (>=0.3.11,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
praw = { version = "^7.8.1", markers = "python_version >= '3.11' and python_version < '4.0'" }
